Add files via upload
Project Description

Term: Fall 2019

Project title: Is the spirit of Hip-Hop means “peace & love “?

This project is conducted by Mo Yang (my2607)

Project summary: We learn some genre of music from books, TV, or internet, we will have an impression of that. However, in real-life, will that type of music the same as what you think about that? My impact of Hip-Hop is to express “love & peace” with an aggressive attitude. The report shows that Hip-Hop has the strongest emotion, and love is the most frequency words. However, there are a lot of dirty words.

I can't output the knit, so I write in a pdf file. Sorry for the inconvenience.